refactor: consolidate LOCALE_MAP to lib/constants.ts (DRY)
- Add shared LOCALE_MAP constant to lib/constants.ts
- Update all 8 tool pages to import from shared location
- Update qr-code/constants.ts to re-export from lib/constants
- Removes ~40 lines of duplicated code across the codebase
