+++ lib/constants.ts
@@ -6,6 +6,17 @@ export const SITE_URL = "https://www.articleideagenerator.com";
 export const SITE_NAME = "Article Idea Generator";
 
 /**
+ * OpenGraph locale mapping for internationalization
+ * Maps app locales to OpenGraph locale format
+ */
+export const LOCALE_MAP: Record<string, string> = {
+  en: "en_US",
+  fr: "fr_FR",
+  es: "es_ES",
+  de: "de_DE",
+};
+
+/**
  * Get locale-aware path
  */
 export function getLocalePath(path: string, locale: string): string {
