+++ pages/tools/cover-image-generator.tsx
@@ -14,7 +14,7 @@ import { Header, Footer } from "@/components";
 import { RelatedTools } from "@/components/tools";
 import { useTranslation } from "@/hooks/useTranslation";
 import { useCoverImage } from "@/hooks/useCoverImage";
-import { SITE_URL, SITE_NAME } from "@/lib/constants";
+import { SITE_URL, SITE_NAME, LOCALE_MAP } from "@/lib/constants";
 import {
   DownloadIcon,
   PhotographIcon,
@@ -77,13 +77,6 @@ import { DEFAULT_IMAGE_FILTERS } from "@/lib/cover-image/editor-types";
 // ============================================================================
 // Constants
 // ============================================================================
-
-const LOCALE_MAP: Record<string, string> = {
-  en: "en_US",
-  fr: "fr_FR",
-  es: "es_ES",
-  de: "de_DE",
-};
 
 const spaceGrotesk = Space_Grotesk({
   weight: "700",
