+++ pages/tools/word-counter.tsx
@@ -7,16 +7,8 @@ import { Header, Footer } from "@/components";
 import { RelatedTools } from "@/components/tools";
 import { StatsPanel, type TextStats } from "@/components/tools";
 import { useTranslation } from "@/hooks/useTranslation";
-import { SITE_URL, SITE_NAME } from "@/lib/constants";
+import { SITE_URL, SITE_NAME, LOCALE_MAP } from "@/lib/constants";
 import { trackToolUsage } from "@/lib/gtag";
-
-// Locale to og:locale format mapping
-const LOCALE_MAP: Record<string, string> = {
-  en: "en_US",
-  fr: "fr_FR",
-  es: "es_ES",
-  de: "de_DE",
-};
 
 const spaceGrotesk = Space_Grotesk({
   weight: "700",
